third_party/WebKit/LayoutTests/inspector/timeline/timeline-time-expected.txt

Test nesting of time/timeEnd records on Timeline


Running: testSimpleConsoleTime
Root
----> FunctionCall
----> FunctionCall
--------> Console Time: a

Running: testNestedConsoleTime
Root
----> FunctionCall
--------> Console Time: a
------------> Console Time: b
------------> Console Time: c
----------------> Console Time: d

Running: testUnbalancedConsoleTime
Root
----> FunctionCall
--------> Console Time: b

Running: testConsoleTimeWithoutConsoleTimeEnd
Root
----> FunctionCall
--------> Timestamp: Foo
--------> Timestamp: Bar
--------> Console Time: d
------------> Timestamp: Baz
